media_modules: CF1 support kernel 6.12 [1/1]
PD#SWPL-188419
PD#SWPL-188420

Problem:
1.strlcpy removed from kernel 6.12
2.usage of vb2_queue num_buffers has changed
3.min_buffers_needed modified
4.platform_driver.remove prototype changed
5.dhp new funcion not compat with kernel 6.6
6.some other compilation issues
7.some spinlock used without init

Solution:
1.use strscpy instead
2.use vb2_get_num_buffers after kernel 6.8
3.use min_reqbufs_allocation after kernel 6.4
4.compat with new prototype
5.compat dhp with kernel 6.6
6.fixing compilation issues
7.initialize these spinlocks

Verify:
adt4

+++ drivers/frame_sink/encoder/vcenc/vc8000_vcmd_driver.c
@@ -3565,7 +3565,7 @@ static int vcmd_pcie_init(struct platform_device *pf_dev)
 	dma_coerce_mask_and_coherent(&pf_dev->dev, DMA_BIT_MASK(64));
 
 	vaddr = dma_alloc_coherent(&pf_dev->dev, alloc_size_byte, &paddr, GFP_KERNEL);
-	pr_info("------- vaddr: %p, paddr: %llx\n", vaddr, paddr);
+	pr_info("------- vaddr: %p, paddr: %llx\n", vaddr, (u64)paddr);
 
 	g_vcmd_base_hdwr = vers_reg_start[VERS_CORE_REG_BASE];
 	pr_info("Base hw val 0x%llx\n", (unsigned long long)g_vcmd_base_hdwr);
